add tests for group, person repository

+++ src/main/java/Database_layer/Repositories/GroupRepository.java
@@ -56,7 +56,10 @@ public class GroupRepository implements IRepository<Group>{
         return groups;
     }
 
-    public Group Get(int id) throws SQLException {
+    public Group Get(int id) throws SQLException, IllegalArgumentException {
+        if (id<1){
+            throw new IllegalArgumentException();
+        }
         Group group = new Group();
         try{
             ResultSet rs = statement.executeQuery("SELECT * FROM `group` WHERE group_id="+id);
@@ -74,7 +77,13 @@ public class GroupRepository implements IRepository<Group>{
         return group;
     }
 
-    public void Add(Group entity) throws SQLException {
+    public void Add(Group entity) throws SQLException,IllegalArgumentException {
+        if (entity == null){
+            throw new IllegalArgumentException();
+        }
+        if(entity.getCourse_id() <1 || entity.getAmount() < 1 || entity.getLevel() == null){
+            throw new IllegalArgumentException();
+        }
         try {
             int resultSet = statement.executeUpdate(
                     String.format("INSERT INTO `group` (level,amount,course_id,person_id) VALUES ('%s','%d','%d','%d')",
@@ -91,7 +100,10 @@ public class GroupRepository implements IRepository<Group>{
         }
     }
 
-    public void Delete(int id) throws SQLException {
+    public void Delete(int id) throws SQLException,IllegalArgumentException {
+        if (id < 1){
+            throw new IllegalArgumentException();
+        }
         try{
             int resultSet = statement.executeUpdate("DELETE FROM `group` WHERE group_id="+id);
             System.out.println("Rows affected during Delete: " + resultSet);
@@ -101,7 +113,13 @@ public class GroupRepository implements IRepository<Group>{
         }
     }
 
-    public void Update(int id, Group item) throws SQLException {
+    public void Update(int id, Group item) throws SQLException, IllegalArgumentException {
+        if (item == null || id<1){
+            throw new IllegalArgumentException();
+        }
+        if(item.getCourse_id() <1 || item.getAmount() < 1 || item.getLevel() == null){
+            throw new IllegalArgumentException();
+        }
         try{
             int resultSet = statement.executeUpdate(String.format("UPDATE `group` SET level='%s',amount='%d',course_id='%d',person_id='%d' WHERE group_id='%s'"
                     ,item.getLevel()
